requirements.txt
Django==4.2.11
djangorestframework==3.14.0
psycopg2-binary==2.9.9
psycopg==3.1.14
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
python-decouple==3.8
pyarrow==15.0.0
pandas==2.2.0
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
freezegun==1.5.1
factory-boy==3.3.0
requests==2.31.0
python-dateutil==2.8.2
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1
django-extensions==3.2.3
python-json-logger==2.0.7

